# swagger.yaml
openapi: 3.0.0
info:
  title: Blog Api
  contact: {}
  version: "1.0"
servers:
  - url: https://blog-api-4.herokuapp.com/api/v1/
    variables: {}
paths:
...
  /comments:
    post:
      tags:
        - Comments
      summary: New Comment
      operationId: NewComment
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
            example:
              text: test
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
    get:
      tags:
        - Comments
      summary: Get All Comment For Specific Article
      operationId: GetAllCommentForSpecificArticle
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
  /comments/{id}:
    delete:
      tags:
        - Comments
      summary: Delete Comment
      operationId: DeleteComment
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
components:
  schemas:
...
    Comment:
      title: Comment
      required:
        - text
      type: object
      properties:
        text:
          type: string
      description: Comment Schema
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...